backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.2.0
ffmpeg-python==0.2.0
celery==5.3.6
redis==5.0.1
drf-spectacular==0.27.0
numpy==1.26.3
librosa==0.10.1
scipy==1.12.0
matplotlib==3.8.2
django-allauth==0.61.1
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
cryptography==42.0.2
